Karma config update, package.json update.
Updated the package.json file with an npm update --save-dev which
changed mostly karma-affiliated packages.

Fixed karma setup that was generated long ago and not used, which caused
the ngResource missing error. Now the config should be up to date and
with 'mocks' support, which will probably be needed e.g. in
MainController.