Rescanning of server-side song directory moved to separate action.
It was irritating to wait for the rescan to complete on each page
refresh. Now on refresh you only request the last result of a scan and
you're able to trigger one in edit mode.

There are plans to enable server side periodic rescan or some kind of
watch mechanism, but for now this is sufficient.