cleanup state management for path, make links optional
Across the $scope there were fields keeping the same exact value, e.g.
listview. They were refactored away, since they were always assigned
together.

To make the location change optional, another wrapper was created, not
to disturb the default behaviour of $location.path call. This wrapper
uses $scope.flags.link to tell if the url should be updated.

Also the attempt to use HTML5 history API to manage the url was
abandoned for now, since it requires server-side url rewriting. And
configuring it in grunt could be troublesome. It'll be a nice-to-have
someday.